refactor: make the wait until ready a common, generic function
Removed from SSH
